Extract HUD into another file for easier reuse
